working on tests

+++ app/src/main/java/com/vhontar/anynotes/business/data/cache/abstraction/NoteCacheDataSource.kt
@@ -6,7 +6,7 @@ interface NoteCacheDataSource {
     suspend fun insertNote(note: Note): Long
     suspend fun deleteNote(primaryKey: String): Int
     suspend fun deleteNotes(notes: List<Note>): Int
-    suspend fun updateNote(primaryKey: String, newTitle: String, newBody: String?, timestamp: String?): Long
+    suspend fun updateNote(primaryKey: String, newTitle: String, newBody: String?, timestamp: String?): Int
     suspend fun searchNotes(query: String, filterAndOrder: String, page: Int): List<Note>
     suspend fun searchNoteById(primaryKey: String): Note?
     suspend fun getNumNotes(): Int
